+++ dune/typetree/traversal.hh
@@ -244,43 +244,6 @@ namespace Dune {
     /**
      * \brief Traverse tree and visit each node
      *
-     * All passed callback functions are called with the
-     * node and corresponding treepath as arguments.
-     *
-     * \param tree The tree to traverse
-     * \param preFunc This function is called for each inner node before visiting its children
-     * \param leafFunc This function is called for each leaf node
-     * \param postFunc This function is called for each inner node after visiting its children
-     *
-     * \deprecated Use the more general \ref applyToTree instead.
-     */
-    template<class Tree, class PreFunc, class LeafFunc, class PostFunc>
-    [[deprecated]] void forEachNode(Tree&& tree, PreFunc&& preFunc, LeafFunc&& leafFunc, PostFunc&& postFunc)
-    {
-      Detail::forEachNode(tree, hybridTreePath(), preFunc, leafFunc, postFunc);
-    }
-
-    /**
-     * \brief Traverse tree and visit each node
-     *
-     * All passed callback functions are called with the
-     * node and corresponding treepath as arguments.
-     *
-     * \param tree The tree to traverse
-     * \param innerFunc This function is called for each inner node before visiting its children
-     * \param leafFunc This function is called for each leaf node
-     *
-     * \deprecated Use the more general \ref applyToTree instead.
-     */
-    template<class Tree, class InnerFunc, class LeafFunc>
-    [[deprecated]] void forEachNode(Tree&& tree, InnerFunc&& innerFunc, LeafFunc&& leafFunc)
-    {
-      Detail::forEachNode(tree, hybridTreePath(), innerFunc, leafFunc, NoOp{});
-    }
-
-    /**
-     * \brief Traverse tree and visit each node
-     *
      * The passed callback function is called with the
      * node and corresponding treepath as arguments.
      *