+++ CHANGELOG.md
@@ -14,6 +14,10 @@ TypeTree 2.10-git
   `max_size`, `join`, `reverse`, and `accumulate_[front|back]`.
 - Add user-defined literal `_tp` to generate `HybridTreePath`
 
+Deprecations and removals
+
+- Remove deprecated `TreePath`, use `StaticTreePath` instead.
+
 TypeTree 2.9
 ----------------