+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Deprecations and removals
 - Remove deprecated `TreePath`, use `StaticTreePath` instead.
 - Remove deprecated `CHILDREN` from `CompositeNode`, `FilteredCompositeNode`, `LeafNode`,
   `PowerNode`, `ProxyNode`. Use `degree()` instead.
+- Remove deprecated `forEachNode`, use `applyToTree` instead.
 
 TypeTree 2.9
 ----------------